ipcache/md: Support incremental ipcache updates
Previously, the InjectLabels() function which reacts to updates to the
IPCache metadata would iterate through the entire metadata map in order
to detect changes that occur, and subsequently react to those updates.
The reaction includes potentially allocating new identities, then
triggering updates into the datapath policymaps and the datapath's ipcache.

Given that there was no way to track which prefixes had been updated,
this meant iterating over the entire map, perhaps re-allocating
identities for prefixes that had not changed. This also made it more
difficult in future to support different types of updates to the ipcache
from other sources, since each feature would need to have special logic
encoded in this function to correctly handle the identity allocation.

Rework the InjectLabels() implementation to handle a queue of prefixes
to update instead. This commit only modifies the path used while
associating new sets of labels with a given prefix. A follow up commit
will handle the delete cases. However, given that adding new sets of
labels could also cause an existing prefix -> Labels (and hence Identity)
mapping to change, even this case must handle the cleanup of old
identities, which is operationally similar to deleting labels (and hence
identities).

This new code is also built around asynchronous updates to the label
associations for the prefixes. When triggered it only tracks which
prefixes have been changed, and not yet what it must do to ensure that
these prefixes have the correct identity. It uses two key pieces of
information to determine what must be done in order to ensure the
correct association:
* The desired state of the labels, prepared in the IPCache.metadata map
  during a previous UpsertMetadata() call, and
* The realized state of the ipcache itself.

The first time that a set of labels is associated with a prefix, the
ipcache will likely not have any corresponding entry, so this case is
simple: Check the desired state of the labels, allocate a corresponding
identity, and then update the policy selectorcache, datapath policymaps,
and finally the ipcache.

The second time that a (new) set of labels is associated with a prefix,
it gets a bit more complicated: The previous set of labels may still be
relevant. First, a lookup into the existing ipcache tells the function
that this prefix is already associated with some set of labels.

* First, determine the complete set of labels that should be associated
  with this prefix. In this commit, the new labels are likely a
  superset of the existing set. A subsequent commit will handle disjoint
  sets.
* Allocate a new identity corresponding to this complete set of labels.
* Update the SelectorCache and endpoint policy maps to potentially allow
  this new identity. Traffic will not yet be identified with this new
  identity, but this prepares the datapath policy engine for that
  eventuality.
* Update the IPcache to identify this prefix as having the new identity.
* Now that the new identity is in use in the datapath, we can remove
  references to the old identity (in the SelectorCache and policymaps)
* Finally, given that the IPCache no longer associates this prefix with
  the old identity, release the reference on the old identity to ensure
  that the old identity is returned back to the identity allocation pool.

Previously, the "identity release" logic was performed as part of the
core loop, and primarily based on whether any allocated identity was
considered "new" or had the "kube-apiserver" labels associated with it,
which was a bit difficult to reason about. That code is now simpler,
being tracked via an `idsToDelete` map during the processing of modified
prefixes, and the release is moved to the end of the function to ensure
that it is not prematurely released.

Co-authored-by: Chris Tarazi <chris@isovalent.com>
Signed-off-by: Joe Stringer <joe@cilium.io>